rename card_to_products.py to sealed.py
