+++ mtgjson5/pipeline/stages/sealed.py
